+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+7.2
+---
+
+ * Use `ChainCache` to store warmed-up cache in `kernel.build_dir` and runtime cache in `kernel.cache_dir`
+ * Make `TemplateCacheWarmer` use `kernel.build_dir` instead of `kernel.cache_dir`
+
 7.1
 ---